+++ vitest.config.ts
@@ -6,9 +6,9 @@ export default defineConfig({
   },
   resolve: {
     alias: {
-      "@nudo/core": new URL("./packages/core/src", import.meta.url)
+      "@nudojs/core": new URL("./packages/core/src", import.meta.url)
         .pathname,
-      "@nudo/parser": new URL("./packages/parser/src", import.meta.url)
+      "@nudojs/parser": new URL("./packages/parser/src", import.meta.url)
         .pathname,
     },
   },